pyproject.toml
[tool.poetry]
name = "file_checker"
version = "1.0.0"
description = "backend"
authors = ["digital.rsue <digital@rsue.ru>"]


[tool.black]
line-length = 79


[tool.isort]
multi_line_output=3
include_trailing_comma=true
use_parentheses=true
line_length=79


[tool.taskipy.tasks]
manage = "python ./manage.py"
clear = "task manage clear"
dumpdata = "task manage dump"
migrate = "task manage migrate --noinput"
server = "task manage runserver"
loaddata = "task manage loaddata"
startapp = "task manage createapp"
initconfig = "task manage initconfig"
makemigrations = "task manage makemigrations"
createsuperuser = "task manage createsuperuser"
collectstatic = "task manage collectstatic --noinput"
gunicorn = "gunicorn --timeout 300 config.wsgi:application -b 0.0.0.0:8000"
defaultadmin = "task manage defaultadmin"
defaultfixtures = "task manage defaultfixtures"


[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.5"
django-cleanup = "^5.1.0"
django-settings-export = "^1.2.1"
djangorestframework = "^3.11.1"
django-cors-headers = "^3.6.0"
drf-yasg = "^1.17.0"
django-import-export = "^2.0.1"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"
mysqlclient = "^2.0.1"
pillow = "^8.1.0"
djangorestframework-camel-case = "^1.1.2"
django-weasyprint = "^1.0.1"
django-service-objects = "^0.7.0"
django-admin-interface = "^0.17.3"
django-extra-settings = "^0.1.3"
dj-rest-auth = "^2.2.3"
djangorestframework-simplejwt = "^5.0.0"
docxtpl = "^0.15.2"
psycopg2 = "^2.9.3"
oletools = "^0.60.1"
pip = "^22.1"
PyPDF2 = "^1.27.12"
pdfid = "^1.1.0"
striprtf = "^0.0.20"
whitenoise = "^6.2.0"
django-filter = "^21.1"
django-crispy-forms = "^1.14.0"


[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
django-debug-toolbar = "^3.1.1"
autopep8 = "^1.5"
django-extensions = "^3.0.9"
pylint-django = "^2.0.13"
taskipy = "^1.1.3"
flake8 = "^3.7.9"
black = "^20.8b1"
Werkzeug = "^1.0.1"
safety = "^1.9.0"
bandit = "^1.7.0"
sphinx = "^3.3.1"
sphinx-rtd-theme = "^0.5.0"
